Working to address https://github.com/jasnell/w3c-socialwg-activitystreams/issues/24#issuec omment-58074056
This does several things:

1. It outlines a clearer separation between as:Object and as:Link
2. It models as:Link in much the same way that hydra:Link is defined,
using it as a marker to indicate that the @id is dereferencable
3. It specifically handles the case where a JSON-LD object is defined
as both an Object and a Link (e.g. "@type": ["as:Link",
"urn:foo:object"])

For anyone looking at this without JSON-LD glasses on, this might look
strange, but it works. Compatibility with AS 1.0 is preserved and
people who want to use the JSON only model (without the additional -LD
stuff) can continue to do so without any additional complication.