adding profile media type parameter
the proposal is to add a profile media type parameter to both media types. it would allow profiles to be identified at the media type level. profiles according to RFC 6906 could be used to signal certain vocabularies that an activity stream may use (or that a client prefers). https://github.com/dret/ASDL is one way how such a profile could be represented (currently ASDL is focusing on AS1), but the profile parameter would not require a specific representation, or any at all. it would simply signify that an activity stream follows a profile.