Initial write-up of Apache Streams
This document reflects the proposed feature set of AS 2.0 which is currently implemented in Apache Streams using AS 1.0

Apache Streams (incubating) supports the efforts of the Social Working Group to publish AS 2.0 as a formal specification.  As more providers are added to Apache Streams and the variety of documents they produce increases, the project intends to align and remain aligned with the AS 2.0 vocubulary.