Make name a MAY not a SHOULD
Per agreement by the WG, the name property is now a MAY, and the
summary property is a fallback for an automatically-generated
text representation of the object.

Added text to this effect to core. Gave guidance on using these
two properties for text representations, and noted how they could
be unreliable.

All examples that had auto-generated-sounding name properties now
use summary.

Closes #312.